[MOD] moved separated all css from js
