+++ src/Users/Admin/AdmHomePage.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { LastAlerts } from '../../Components/Alerts/LastAlerts'
 import HeaderComp from '../../Components/Header/HeaderComp'
 import Sidebar from '../../Components/Sidebar/Sidebar'
-import './AdmHomePage.css'
+import '../../css/Users/Admin/AdmHomePage.css'
 
 export default function AdmHomePage () {
   return (
