+++ src/Users/SchoolAdmin/SchoolAdmAccountsPage.js
@@ -2,7 +2,7 @@ import React from 'react'
 import HeaderComp from '../../Components/Header/HeaderComp'
 import Sidebar from '../../Components/Sidebar/Sidebar'
 import SchoolAccountsTable from '../../Components/Accounts/SchoolAdm/SchoolAccountsTable'
-import './SchoolAdmAccountsPage.css'
+import '../../css/Users/SchoolAdmin/SchoolAdmAccountsPage.css'
 
 export default function SchoolAdmAccountsPage () {
   return (
