+++ src/Users/SchoolAdmin/SchoolAdmHomePage.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { LastAlerts } from '../../Components/Alerts/LastAlerts'
 import HeaderComp from '../../Components/Header/HeaderComp'
 import Sidebar from '../../Components/Sidebar/Sidebar'
-import './SchoolAdmHomePage.css'
+import '../../css/Users/SchoolAdmin/SchoolAdmHomePage.css'
 import { QuestSpace } from '../../Components/Questionnaire/QuestSpace'
 import { GraphSpace } from '../../Components/Graph/GraphSpace'
